+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file. For commit guidelines, please refer to [Standard Version](https://github.com/conventional-changelog/standard-version).
 
+## v0.3.6
+
+  **Notes**:
+    - All views modes to show header bar for sorting.
+    - slight style changes
+
+  **Bugfixes**:
+    - select and info bug after sorting https://github.com/gtsteffaniak/filebrowser/issues/277
+
 ## v0.3.5
 
   **New Features**